+++ pages/api/adoptionpost/type/[breed].ts
@@ -1,4 +1,4 @@
-import { prisma } from "../../../../../lib/prisma";
+import { prisma } from "../../../../lib/prisma";
 import { NextApiRequest, NextApiResponse } from "next";
 
 export default async function handler(
