--- pages/api/create/professional.ts
@@ -1,40 +1,0 @@
-import { prisma } from "../../../lib/prisma";
-import { NextApiRequest, NextApiResponse } from "next";
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  const {
-    businessName,
-    contact,
-    address,
-    description,
-    photo,
-    dni,
-    type,
-    qualification,
-    active,
-    id,
-  } = req.body;
-  try {
-    await prisma.professionalBusiness.create({
-      data: {
-        businessName,
-        contact,
-        address,
-        description,
-        photo,
-        dni,
-        type,
-        active,
-        qualification,
-        ownerBusinessId: id,
-      },
-    });
-    res.status(200).json({ message: "created " });
-  } catch (error) {
-    console.log(error);
-    res.status(400).json({ message: "error" });
-  }
-}
